+++ flexmeasures/data/models/planning/__init__.py
@@ -173,6 +173,19 @@ class Scheduler:
         """
         pass
 
+    def collect_flex_config(self):
+        """Merge the flex-config from the db (from the asset and its ancestors) with the initialization flex-config.
+
+        Note that self.flex_context overrides db_flex_context (from the asset and its ancestors).
+        # todo: also fetch db_flex_model once that exists
+        """
+        if self.asset is not None:
+            asset = self.asset
+        else:
+            asset = self.sensor.generic_asset
+        db_flex_context = asset.get_flex_context()
+        self.flex_context = {**db_flex_context, **self.flex_context}
+
     def deserialize_config(self):
         """
         Check all configurations we have, throwing either ValidationErrors or ValueErrors.