+++ flexmeasures/data/models/planning/storage.py
@@ -163,9 +163,7 @@ class MetaStorageScheduler(Scheduler):
         up_deviation_prices = get_continuous_series_sensor_or_quantity(
             variable_quantity=consumption_price,
             actuator=asset,
-            unit=FlexContextSchema()
-            .declared_fields["consumption_price"]
-            ._get_unit(consumption_price),
+            unit=self.flex_context["shared_currency_unit"] + "/MWh",
             query_window=(start, end),
             resolution=resolution,
             beliefs_before=belief_time,
@@ -175,9 +173,7 @@ class MetaStorageScheduler(Scheduler):
         down_deviation_prices = get_continuous_series_sensor_or_quantity(
             variable_quantity=production_price,
             actuator=asset,
-            unit=FlexContextSchema()
-            .declared_fields["production_price"]
-            ._get_unit(production_price),
+            unit=self.flex_context["shared_currency_unit"] + "/MWh",
             query_window=(start, end),
             resolution=resolution,
             beliefs_before=belief_time,
@@ -276,9 +272,7 @@ class MetaStorageScheduler(Scheduler):
             ems_peak_consumption_price = get_continuous_series_sensor_or_quantity(
                 variable_quantity=ems_peak_consumption_price,
                 actuator=asset,
-                unit=FlexContextSchema()
-                .declared_fields["ems_peak_consumption_price"]
-                ._get_unit(ems_peak_consumption_price),
+                unit=self.flex_context["shared_currency_unit"] + "/MW",
                 query_window=(start, end),
                 resolution=resolution,
                 beliefs_before=belief_time,
@@ -312,9 +306,7 @@ class MetaStorageScheduler(Scheduler):
             ems_peak_production_price = get_continuous_series_sensor_or_quantity(
                 variable_quantity=ems_peak_production_price,
                 actuator=asset,
-                unit=FlexContextSchema()
-                .declared_fields["ems_peak_production_price"]
-                ._get_unit(ems_peak_production_price),
+                unit=self.flex_context["shared_currency_unit"] + "/MW",
                 query_window=(start, end),
                 resolution=resolution,
                 beliefs_before=belief_time,
@@ -350,9 +342,7 @@ class MetaStorageScheduler(Scheduler):
             any_ems_consumption_breach_price = get_continuous_series_sensor_or_quantity(
                 variable_quantity=ems_consumption_breach_price,
                 actuator=asset,
-                unit=FlexContextSchema()
-                .declared_fields["ems_consumption_breach_price"]
-                ._get_unit(ems_consumption_breach_price),
+                unit=self.flex_context["shared_currency_unit"] + "/MW",
                 query_window=(start, end),
                 resolution=resolution,
                 beliefs_before=belief_time,
@@ -361,10 +351,8 @@ class MetaStorageScheduler(Scheduler):
             all_ems_consumption_breach_price = get_continuous_series_sensor_or_quantity(
                 variable_quantity=ems_consumption_breach_price,
                 actuator=asset,
-                unit=FlexContextSchema()
-                .declared_fields["ems_consumption_breach_price"]
-                ._get_unit(ems_consumption_breach_price)
-                + "*h",  # from EUR/MWh to EUR/MW/resolution
+                unit=self.flex_context["shared_currency_unit"]
+                + "/MW*h",  # from EUR/MWh to EUR/MW/resolution
                 query_window=(start, end),
                 resolution=resolution,
                 beliefs_before=belief_time,
@@ -404,9 +392,7 @@ class MetaStorageScheduler(Scheduler):
             any_ems_production_breach_price = get_continuous_series_sensor_or_quantity(
                 variable_quantity=ems_production_breach_price,
                 actuator=asset,
-                unit=FlexContextSchema()
-                .declared_fields["ems_production_breach_price"]
-                ._get_unit(ems_production_breach_price),
+                unit=self.flex_context["shared_currency_unit"] + "/MW",
                 query_window=(start, end),
                 resolution=resolution,
                 beliefs_before=belief_time,
@@ -415,10 +401,8 @@ class MetaStorageScheduler(Scheduler):
             all_ems_production_breach_price = get_continuous_series_sensor_or_quantity(
                 variable_quantity=ems_production_breach_price,
                 actuator=asset,
-                unit=FlexContextSchema()
-                .declared_fields["ems_production_breach_price"]
-                ._get_unit(ems_production_breach_price)
-                + "*h",  # from EUR/MWh to EUR/MW/resolution
+                unit=self.flex_context["shared_currency_unit"]
+                + "/MW*h",  # from EUR/MWh to EUR/MW/resolution
                 query_window=(start, end),
                 resolution=resolution,
                 beliefs_before=belief_time,
@@ -525,9 +509,7 @@ class MetaStorageScheduler(Scheduler):
                 any_soc_minima_breach_price = get_continuous_series_sensor_or_quantity(
                     variable_quantity=soc_minima_breach_price,
                     actuator=asset,
-                    unit=FlexContextSchema()
-                    .declared_fields["soc_minima_breach_price"]
-                    ._get_unit(soc_minima_breach_price),
+                    unit=self.flex_context["shared_currency_unit"] + "/MWh",
                     query_window=(start + resolution, end + resolution),
                     resolution=resolution,
                     beliefs_before=belief_time,
@@ -537,10 +519,8 @@ class MetaStorageScheduler(Scheduler):
                 all_soc_minima_breach_price = get_continuous_series_sensor_or_quantity(
                     variable_quantity=soc_minima_breach_price,
                     actuator=asset,
-                    unit=FlexContextSchema()
-                    .declared_fields["soc_minima_breach_price"]
-                    ._get_unit(soc_minima_breach_price)
-                    + "*h",  # from EUR/MWh² to EUR/MWh/resolution
+                    unit=self.flex_context["shared_currency_unit"]
+                    + "/MWh*h",  # from EUR/MWh² to EUR/MWh/resolution
                     query_window=(start + resolution, end + resolution),
                     resolution=resolution,
                     beliefs_before=belief_time,
@@ -609,9 +589,7 @@ class MetaStorageScheduler(Scheduler):
                 any_soc_maxima_breach_price = get_continuous_series_sensor_or_quantity(
                     variable_quantity=soc_maxima_breach_price,
                     actuator=asset,
-                    unit=FlexContextSchema()
-                    .declared_fields["soc_maxima_breach_price"]
-                    ._get_unit(soc_maxima_breach_price),
+                    unit=self.flex_context["shared_currency_unit"] + "/MWh",
                     query_window=(start + resolution, end + resolution),
                     resolution=resolution,
                     beliefs_before=belief_time,
@@ -621,10 +599,8 @@ class MetaStorageScheduler(Scheduler):
                 all_soc_maxima_breach_price = get_continuous_series_sensor_or_quantity(
                     variable_quantity=soc_maxima_breach_price,
                     actuator=asset,
-                    unit=FlexContextSchema()
-                    .declared_fields["soc_maxima_breach_price"]
-                    ._get_unit(soc_maxima_breach_price)
-                    + "*h",  # from EUR/MWh² to EUR/MWh/resolution
+                    unit=self.flex_context["shared_currency_unit"]
+                    + "/MWh*h",  # from EUR/MWh² to EUR/MWh/resolution
                     query_window=(start + resolution, end + resolution),
                     resolution=resolution,
                     beliefs_before=belief_time,
@@ -730,9 +706,7 @@ class MetaStorageScheduler(Scheduler):
                         get_continuous_series_sensor_or_quantity(
                             variable_quantity=production_breach_price,
                             actuator=asset,
-                            unit=FlexContextSchema()
-                            .declared_fields["production_breach_price"]
-                            ._get_unit(production_breach_price),
+                            unit=self.flex_context["shared_currency_unit"] + "/MW",
                             query_window=(start, end),
                             resolution=resolution,
                             beliefs_before=belief_time,
@@ -744,10 +718,8 @@ class MetaStorageScheduler(Scheduler):
                         get_continuous_series_sensor_or_quantity(
                             variable_quantity=production_breach_price,
                             actuator=asset,
-                            unit=FlexContextSchema()
-                            .declared_fields["production_breach_price"]
-                            ._get_unit(production_breach_price)
-                            + "*h",  # from EUR/MWh to EUR/MW/resolution
+                            unit=self.flex_context["shared_currency_unit"]
+                            + "/MW*h",  # from EUR/MWh to EUR/MW/resolution
                             query_window=(start, end),
                             resolution=resolution,
                             beliefs_before=belief_time,
@@ -806,9 +778,7 @@ class MetaStorageScheduler(Scheduler):
                         get_continuous_series_sensor_or_quantity(
                             variable_quantity=consumption_breach_price,
                             actuator=asset,
-                            unit=FlexContextSchema()
-                            .declared_fields["consumption_breach_price"]
-                            ._get_unit(consumption_breach_price),
+                            unit=self.flex_context["shared_currency_unit"] + "/MW",
                             query_window=(start, end),
                             resolution=resolution,
                             beliefs_before=belief_time,
@@ -820,10 +790,8 @@ class MetaStorageScheduler(Scheduler):
                         get_continuous_series_sensor_or_quantity(
                             variable_quantity=consumption_breach_price,
                             actuator=asset,
-                            unit=FlexContextSchema()
-                            .declared_fields["consumption_breach_price"]
-                            ._get_unit(consumption_breach_price)
-                            + "*h",  # from EUR/MWh to EUR/MW/resolution
+                            unit=self.flex_context["shared_currency_unit"]
+                            + "/MW*h",  # from EUR/MWh to EUR/MW/resolution
                             query_window=(start, end),
                             resolution=resolution,
                             beliefs_before=belief_time,
@@ -889,24 +857,32 @@ class MetaStorageScheduler(Scheduler):
                 device_constraints[d]["stock delta"] *= timedelta(hours=1) / resolution
 
             # Apply round-trip efficiency evenly to charging and discharging
-            charging_efficiency[d] = get_continuous_series_sensor_or_quantity(
-                variable_quantity=charging_efficiency[d],
-                actuator=sensor_d,
-                unit="dimensionless",
-                query_window=(start, end),
-                resolution=resolution,
-                beliefs_before=belief_time,
-                fallback_attribute="charging-efficiency",
-            ).fillna(1)
-            discharging_efficiency[d] = get_continuous_series_sensor_or_quantity(
-                variable_quantity=discharging_efficiency[d],
-                actuator=sensor_d,
-                unit="dimensionless",
-                query_window=(start, end),
-                resolution=resolution,
-                beliefs_before=belief_time,
-                fallback_attribute="discharging-efficiency",
-            ).fillna(1)
+            charging_efficiency[d] = (
+                get_continuous_series_sensor_or_quantity(
+                    variable_quantity=charging_efficiency[d],
+                    actuator=sensor_d,
+                    unit="dimensionless",
+                    query_window=(start, end),
+                    resolution=resolution,
+                    beliefs_before=belief_time,
+                    fallback_attribute="charging-efficiency",
+                )
+                .astype(float)
+                .fillna(1)
+            )
+            discharging_efficiency[d] = (
+                get_continuous_series_sensor_or_quantity(
+                    variable_quantity=discharging_efficiency[d],
+                    actuator=sensor_d,
+                    unit="dimensionless",
+                    query_window=(start, end),
+                    resolution=resolution,
+                    beliefs_before=belief_time,
+                    fallback_attribute="discharging-efficiency",
+                )
+                .astype(float)
+                .fillna(1)
+            )
 
             roundtrip_efficiency = flex_model[d].get(
                 "roundtrip_efficiency",
@@ -940,6 +916,7 @@ class MetaStorageScheduler(Scheduler):
                         fallback_attribute="storage_efficiency",  # this should become storage-efficiency
                         max_value=1,
                     )
+                    .astype(float)
                     .fillna(1.0)
                     .clip(lower=0.0, upper=1.0)
                 )
@@ -1004,15 +981,8 @@ class MetaStorageScheduler(Scheduler):
         if self.flex_model is None:
             self.flex_model = {}
 
-        # self.flex_context overrides db_flex_context (from the asset and its ancestors)
-        if self.asset is not None:
-            asset = self.asset
-        else:
-            asset = self.sensor.generic_asset
-        db_flex_context = asset.get_flex_context()
-        self.flex_context = FlexContextSchema().load(
-            {**db_flex_context, **self.flex_context}
-        )
+        self.collect_flex_config()
+        self.flex_context = FlexContextSchema().load(self.flex_context)
 
         if isinstance(self.flex_model, dict):
             # Check state of charge.
@@ -1338,7 +1308,9 @@ class StorageScheduler(MetaStorageScheduler):
                     / timedelta(hours=1),
                     up_efficiency=device_constraints[d]["derivative up efficiency"],
                     down_efficiency=device_constraints[d]["derivative down efficiency"],
-                    storage_efficiency=device_constraints[d]["efficiency"].fillna(1),
+                    storage_efficiency=device_constraints[d]["efficiency"]
+                    .astype(float)
+                    .fillna(1),
                 ),
                 from_unit="MWh",
                 to_unit=flex_model_d["state_of_charge"].unit,
@@ -1557,8 +1529,8 @@ def add_storage_constraints(
             resolution,
         )
 
-    storage_device_constraints["min"] = storage_device_constraints["min"].fillna(
-        soc_min_change
+    storage_device_constraints["min"] = (
+        storage_device_constraints["min"].astype(float).fillna(soc_min_change)
     )
 
     if soc_maxima is not None:
@@ -1570,8 +1542,8 @@ def add_storage_constraints(
             resolution,
         )
 
-    storage_device_constraints["max"] = storage_device_constraints["max"].fillna(
-        soc_max_change
+    storage_device_constraints["max"] = (
+        storage_device_constraints["max"].astype(float).fillna(soc_max_change)
     )
 
     # limiting max and min to be in the range [soc_min, soc_max]
@@ -1787,8 +1759,18 @@ def validate_constraint(
 
     columns_involved = columns_lhs + columns_rhs
 
-    lhs = constraints_df.fillna(0).eval(lhs_expression).round(round_to_decimals)
-    rhs = constraints_df.fillna(0).eval(rhs_expression).round(round_to_decimals)
+    lhs = (
+        constraints_df.astype(float)
+        .fillna(0)
+        .eval(lhs_expression)
+        .round(round_to_decimals)
+    )
+    rhs = (
+        constraints_df.astype(float)
+        .fillna(0)
+        .eval(rhs_expression)
+        .round(round_to_decimals)
+    )
 
     condition = None