+++ flexmeasures/data/schemas/sensors.py
@@ -95,7 +95,7 @@ class TimedEventSchema(Schema):
         setattr(self.fields["value"], "return_magnitude", return_magnitude)
 
     @validates("value")
-    def validate_value(self, _value):
+    def validate_value(self, _value, **kwargs):
         if self.value_validator is not None:
             self.value_validator(_value)
 
@@ -179,7 +179,7 @@ class SensorSchemaMixin(Schema):
     attributes = JSON(required=False)
 
     @validates("unit")
-    def validate_unit(self, unit: str):
+    def validate_unit(self, unit: str, **kwargs):
         if not is_valid_unit(unit):
             raise ValidationError(f"Unit '{unit}' cannot be handled.")
 
@@ -192,7 +192,7 @@ class SensorSchema(SensorSchemaMixin, ma.SQLAlchemySchema):
     generic_asset_id = fields.Integer(required=True)
 
     @validates("generic_asset_id")
-    def validate_generic_asset(self, generic_asset_id: int):
+    def validate_generic_asset(self, generic_asset_id: int, **kwargs):
         generic_asset = db.session.get(GenericAsset, generic_asset_id)
         if not generic_asset:
             raise ValidationError(
@@ -238,7 +238,7 @@ class SensorIdField(MarshmallowClickMixin, fields.Int):
         sensor = db.session.get(Sensor, value)
 
         if sensor is None:
-            raise FMValidationError(f"No sensor found with id {value}.")
+            raise FMValidationError(f"No sensor found with ID {value}.")
 
         # lazy loading now (sensor is somehow not in session after this)
         sensor.generic_asset
@@ -469,7 +469,7 @@ class TimeSeriesOrSensor(VariableQuantityField):
 
 class SensorDataFileSchema(Schema):
     uploaded_files = fields.List(
-        fields.Field(metadata={"type": "string", "format": "byte"}),
+        fields.Raw(metadata={"type": "file"}),
         data_key="uploaded-files",
     )
     belief_time_measured_instantly = fields.Boolean(
@@ -549,7 +549,7 @@ class SensorDataFileSchema(Schema):
                         pd.Timedelta(days=0) if belief_time_measured_instantly else None
                     ),
                     resample=(
-                        True if sensor.event_resolution is not timedelta(0) else False
+                        True if sensor.event_resolution != timedelta(0) else False
                     ),
                     timezone=sensor.timezone,
                 )
@@ -568,4 +568,5 @@ class SensorDataFileSchema(Schema):
         if errors:
             raise ValidationError(errors)
         fields["data"] = dfs
+        fields["filenames"] = [file.filename for file in files]
         return fields